return 403 if you try to delete another account

+++ src/service/user_service.py
@@ -147,10 +147,13 @@ class UserService:
             return internal_err_resp()
 
     @staticmethod
-    def delete_account(user_uuid):
+    def delete_account(user_uuid, connected_user_uuid):
         """" Delete user account """
         if not (user := UserModel.query.filter_by(uuid=user_uuid).first()):
             return err_resp("User not found!", 404)
+
+        if user_uuid != connected_user_uuid:
+            return err_resp("Unable to delete an account which is not your's", 403)
 
         try:
             UserModel.query.filter_by(uuid=user_uuid).delete()
